feat: admin password change from UI. fixes #228 fixes #118
