+++ api/test/ledger.test.js
@@ -2,7 +2,7 @@
 
 // Due to https://github.com/interledgerjs/ilp-kit/blob/master/bin/normalizeEnv.js#L24
 // this will end up loading ../../test.env.list if process.env.NODE_ENV === 'test':
-require('../../bin/normalizeEnv')
+require('../../bin/env').normalizeEnv()
 
 const constitute = require('constitute')
 const Ledger = require('../src/lib/ledger')
