+++ bin/ledger.js
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
-require('./normalizeEnv')
+require('./env').normalizeEnv()
 
 if (process.env.LEDGER_ENABLE) {
   require('five-bells-ledger/src/services/app').start()
